backend/requirements.txt
Flask==2.3.3
Flask-CORS==4.0.0
Flask-JWT-Extended==4.6.0
Flask-SQLAlchemy==3.1.1
bcrypt==4.1.2
chromadb==0.4.15
sentence-transformers>=2.7.0
openai>=1.3.0
pdfplumber==0.10.3
python-docx==1.1.0
python-pptx==0.6.23
tiktoken==0.5.1
python-dotenv==1.0.0
werkzeug==2.3.7
huggingface-hub>=0.17.0
